+++ src/autogluon/cloud/backend/multiodal_sagemaker_backend.py
@@ -1,3 +1,5 @@
+import copy
+import os
 from typing import Optional, Union
 
 import pandas as pd
@@ -137,3 +139,115 @@ class MultiModalSagemakerBackend(SagemakerBackend):
             return pred
 
         return proba
+
+    def predict(
+        self,
+        test_data: Union[str, pd.DataFrame],
+        test_data_image_column: Optional[str] = None,
+        **kwargs,
+    ) -> Optional[pd.Series]:
+        """
+        Predict using SageMaker batch transform.
+        When minimizing latency isn't a concern, then the batch transform functionality may be easier, more scalable, and more appropriate.
+        If you want to minimize latency, use `predict_real_time()` instead.
+        To learn more: https://docs.aws.amazon.com/sagemaker/latest/dg/batch-transform.html
+        This method would first create a AutoGluonSagemakerInferenceModel with the trained predictor,
+        then create a transformer with it, and call transform in the end.
+
+        Parameters
+        ----------
+        test_data: str
+            The test data to be inferenced.
+            Can be a pandas.DataFrame or a local path to a csv file.
+            When predicting multimodality with image modality:
+                You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
+            When predicting with only images:
+                Can be a local path to a directory containing the images or a local path to a single image.
+        test_data_image_column: Optional(str)
+            If test_data involves image modality, you must specify the column name corresponding to image paths.
+            The path MUST be an abspath
+        kwargs:
+            Refer to `SagemakerBackend.predict()`
+        """
+        image_modality_only = self._check_image_modality_only(test_data)
+
+        if image_modality_only:
+            processed_args = self._prepare_image_predict_args(**kwargs)
+            return super().predict(
+                test_data,
+                test_data_image_column=None,
+                transformer_kwargs=processed_args["transformer_kwargs"],
+                transform_kwargs=processed_args["transform_kwargs"] ** kwargs,
+            )
+        else:
+            return super().predict(
+                test_data,
+                test_data_image_column=test_data_image_column,
+                **kwargs,
+            )
+
+    def predict_proba(
+        self,
+        test_data: Union[str, pd.DataFrame],
+        test_data_image_column: Optional[str] = None,
+        **kwargs,
+    ) -> Optional[pd.Series]:
+        """
+        Predict proba using SageMaker batch transform.
+        When minimizing latency isn't a concern, then the batch transform functionality may be easier, more scalable, and more appropriate.
+        If you want to minimize latency, use `predict_real_time()` instead.
+        To learn more: https://docs.aws.amazon.com/sagemaker/latest/dg/batch-transform.html
+        This method would first create a AutoGluonSagemakerInferenceModel with the trained predictor,
+        then create a transformer with it, and call transform in the end.
+
+        Parameters
+        ----------
+        test_data: str
+            The test data to be inferenced.
+            Can be a pandas.DataFrame or a local path to a csv file.
+            When predicting multimodality with image modality:
+                You need to specify `test_data_image_column`, and make sure the image column contains relative path to the image.
+            When predicting with only images:
+                Can be a local path to a directory containing the images or a local path to a single image.
+        test_data_image_column: Optional(str)
+            If test_data involves image modality, you must specify the column name corresponding to image paths.
+            The path MUST be an abspath
+        kwargs:
+            Refer to `SagemakerBackend.predict_proba()`
+        """
+        image_modality_only = self._check_image_modality_only(test_data)
+
+        if image_modality_only:
+            processed_args = self._prepare_image_predict_args(**kwargs)
+            return super().predict_proba(
+                test_data,
+                test_data_image_column=None,
+                transformer_kwargs=processed_args["transformer_kwargs"],
+                transform_kwargs=processed_args["transform_kwargs"] ** kwargs,
+            )
+        else:
+            return super().predict_proba(
+                test_data,
+                test_data_image_column=test_data_image_column,
+                **kwargs,
+            )
+
+    def _prepare_image_predict_args(self, **predict_kwargs):
+        split_type = None
+        content_type = "application/x-image"
+        predict_kwargs = copy.deepcopy(predict_kwargs)
+        transformer_kwargs = predict_kwargs.pop("transformer_kwargs", {})
+        transformer_kwargs["strategy"] = "SingleRecord"
+        transform_kwargs = predict_kwargs.pop("transofrm_kwargs", {})
+        transform_kwargs["split_type"] = split_type
+        transform_kwargs["content_type"] = content_type
+
+        return {"transformer_kwargs": transformer_kwargs, "transform_kwargs": transform_kwargs}
+
+    def _check_image_modality_only(self, test_data):
+        image_modality_only = False
+        if isinstance(test_data, str):
+            if os.path.isdir(test_data) or is_image_file(test_data):
+                image_modality_only = True
+
+        return image_modality_only
