+++ src/autogluon/cloud/endpoint/endpoint.py
@@ -6,6 +6,12 @@ import pandas as pd
 
 class Endpoint(ABC):
     @abstractmethod
+    @property
+    def endpoint_name(self) -> str:
+        """Name of the endpoint"""
+        raise NotImplementedError
+
+    @abstractmethod
     def predict(self, test_data: Union[str, pd.DataFrame], **kwargs) -> Union[pd.DataFrame, pd.Series]:
         """
         Predict with the endpoint
